Sepoilcy :  add platform_app rules for enabling rw access for VT calls
CSVT app need rw access on smd11 device node.
As this is oem app which will not possible to have a domain and
will run as  platform_app.
adding rule to allow rw access to smd11.
Since ShareUid of Phone is removed to avoid 32/64 bit ABI issue,
CSVT is not able to make VT calls.

CRs-Fixed: 815598

Change-Id: If6be8f8af0c675f4e50ed4aaaf103044bc3dc9bf