sepolicy: Allow IMS read access to netmgrd stat file.
IMS is client of netmgrd which needs to have unicast
netlink message communication with netmgrd. For this,
it needs to obtain the pid of the netmgrd. To determine the pid of
netmgrd, IMS needs access to parse through the proc file system and
find the entry corresponding to netmgrd.

Change-Id: Ib799cdc7cbb8a2b9bf4145a58ce2e8772fa105eb
CRs-Fixed: 824248