DO NOT MERGE: Common: SEPolicy: Boot Optimization
Some regex match a lot of directories even when they
do not have to set any  se label on them.
They end up taking a lot of time and slow down
the boot process. This commit removes one such regex
and hardcodes the path to the labels which it
was setting.

Change-Id: Ic42576778c0042f3df521cc2117ab236b7f36a39