sepolicy: Remove getprop permissions for default_prop
Do not allow vendor_init to do read operations on default_prop
as this leads to a compilation failure. This permission was added to
support the read of sys.boot_completed and init.svc.bootanim prop.
However, in Android S, none of these properties are labelled as
default prop and thus it is safe to remove this permission.

Change-Id: I4f7427e883416f619c1882ae38c5d5d9b043e692